+++ packages/blade/src/components/Input/BaseInput/BaseInputTagSlot.web.tsx
@@ -9,6 +9,7 @@ import { useIsomorphicLayoutEffect } from '~utils/useIsomorphicLayoutEffect';
 import { useIsMobile } from '~utils/useIsMobile';
 import { MetaConstants } from '~utils/metaAttribute';
 import { size as sizeToken } from '~tokens/global';
+import { useTableEditableCell } from '~components/Table/TableEditableCellContext';
 
 const MINUMUM_INPUT_SPACE = 60;
 const PLUS_X_MORE_TEXT_WIDTH = 60;
@@ -31,9 +32,10 @@ const useVisibleTagsCount = ({
 }): number => {
   const [visibleTagsCount, setVisibleTagsCount] = React.useState(0);
   const visibleTagsCountStateRef = React.useRef<number>(0);
+  const { isInsideTableEditableCell } = useTableEditableCell();
 
   useIsomorphicLayoutEffect(() => {
-    if (!tags || labelPrefix) {
+    if (!tags || labelPrefix || isInsideTableEditableCell) {
       setVisibleTagsCount(0);
       return;
     }
@@ -109,6 +111,30 @@ const TagSlotContainer = styled(BaseBox)(() => {
   };
 });
 
+const SelectedCountText = ({
+  children,
+  isDisabled,
+}: {
+  children: string;
+  isDisabled?: boolean;
+}): React.ReactElement => {
+  return (
+    <Text
+      color={isDisabled ? 'surface.text.gray.disabled' : 'surface.text.gray.subtle'}
+      alignSelf="center"
+      marginY="spacing.2"
+      marginRight="spacing.4"
+      variant="body"
+      size="small"
+      weight="regular"
+    >
+      <BaseBox as="span" whiteSpace="nowrap">
+        {children}
+      </BaseBox>
+    </Text>
+  );
+};
+
 const BaseInputTagSlot = ({
   renderAs,
   children,
@@ -127,6 +153,7 @@ const BaseInputTagSlot = ({
 }: BaseInputTagSlotProps): React.ReactElement => {
   const hasTags = tags && tags.length > 0;
   const slotRef = React.useRef<HTMLDivElement>(null);
+  const { isInsideTableEditableCell } = useTableEditableCell();
   const visibleTagsCount = useVisibleTagsCount({
     slotRef,
     tags,
@@ -200,7 +227,8 @@ const BaseInputTagSlot = ({
       minHeight={makeSize(baseInputHeight[size])}
       maxHeight={
         // In TextArea with tagged input, we explicitly define maxHeight based on maxHeight so that tags dont overflow out of textarea
-        isDropdownTrigger && isTextArea
+        // And In table we strictly want the maxHeight to be defined to not mess up the table layout
+        (isDropdownTrigger && isTextArea) || isInsideTableEditableCell
           ? makeSize(baseInputHeight[size] * (numberOfLines ?? 1))
           : undefined
       }
@@ -214,30 +242,32 @@ const BaseInputTagSlot = ({
         setShouldIgnoreBlurAnimation?.(false);
       }}
     >
-      {visibleTags}
-      {tags && !showAllTags && invisibleTagsCount ? (
-        <Text
-          color={isDisabled ? 'surface.text.gray.disabled' : 'surface.text.gray.subtle'}
-          alignSelf="center"
-          marginY="spacing.2"
-          marginRight="spacing.4"
-          variant="body"
-          size="small"
-          weight="regular"
-        >
-          <BaseBox as="span" whiteSpace="nowrap">
-            {visibleTags?.length === 0
-              ? getSelectedTextWithoutTags({
-                  items: invisibleTagsCount,
-                  labelPrefix,
-                })
-              : `+${invisibleTagsCount} More`}
-          </BaseBox>
-        </Text>
-      ) : null}
+      {isInsideTableEditableCell && tags && tags.length > 0 ? (
+        <SelectedCountText isDisabled={isDisabled}>
+          {getSelectedTextWithoutTags({ items: tags.length, labelPrefix })}
+        </SelectedCountText>
+      ) : (
+        <>
+          {visibleTags}
+          {tags && !showAllTags && invisibleTagsCount ? (
+            <SelectedCountText isDisabled={isDisabled}>
+              {visibleTags?.length === 0
+                ? getSelectedTextWithoutTags({
+                    items: invisibleTagsCount,
+                    labelPrefix,
+                  })
+                : `+${invisibleTagsCount} More`}
+            </SelectedCountText>
+          ) : null}
+        </>
+      )}
       <BaseBox
         marginTop="-4px"
-        minWidth={hasTags && renderAs === 'button' ? undefined : makeSize(MINUMUM_INPUT_SPACE)}
+        minWidth={
+          hasTags && renderAs === 'button'
+            ? undefined
+            : `min(20%, ${makeSize(MINUMUM_INPUT_SPACE)})`
+        }
         width={hasTags && renderAs === 'button' ? makeSize(sizeToken['1']) : '100%'}
       >
         {children}