+++ packages/blade/src/components/Input/DropdownInputTriggers/BaseDropdownInputTrigger.tsx
@@ -11,6 +11,13 @@ import { MetaConstants } from '~utils/metaAttribute';
 import { getTagsGroup } from '~components/Tag/getTagsGroup';
 import type { BladeElementRef } from '~utils/types';
 import { useFirstRender } from '~utils/useFirstRender';
+import { useTableContext } from '~components/Table/TableContext';
+import {
+  rowDensityToIsTableInputCellMapping,
+  tableEditableCellRowDensityToInputSizeMap,
+  validationStateToInputTrailingIconMap,
+} from '~components/Table/tokens';
+import { useTableEditableCell } from '~components/Table/TableEditableCellContext';
 
 const useControlledDropdownInput = (
   props: Pick<
@@ -142,6 +149,8 @@ const _BaseDropdownInputTrigger = (
     setChangeCallbackTriggerer,
     changeCallbackTriggerer,
   } = useDropdown();
+  const { rowDensity } = useTableContext();
+  const { isInsideTableEditableCell } = useTableEditableCell();
 
   const dropdownTriggerPlaceholder = props.placeholder ?? 'Select Option';
   const isAutoCompleteInHeader = !props.isSelectInput && hasAutoCompleteInBottomSheetHeader;
@@ -216,6 +225,17 @@ const _BaseDropdownInputTrigger = (
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [selectedIndices, selectionType, activeTagIndex, changeCallbackTriggerer, options],
   );
+
+  const tableInputProps: Partial<BaseInputProps> = {
+    isTableInputCell: rowDensityToIsTableInputCellMapping[rowDensity],
+    id: 'table-editable-cell-input',
+    size: tableEditableCellRowDensityToInputSizeMap[rowDensity],
+    trailingIcon: validationStateToInputTrailingIconMap[props.validationState ?? 'none'],
+    showHintsAsTooltip: true,
+  };
+
+  const isValidationStateNone =
+    props.validationState === 'none' || props.validationState === undefined;
 
   return (
     <BaseInput
@@ -297,7 +317,7 @@ const _BaseDropdownInputTrigger = (
       onKeyDown={props.onTriggerKeydown}
       size={props.size}
       trailingInteractionElement={
-        isAutoCompleteInHeader ? null : (
+        isAutoCompleteInHeader || (isInsideTableEditableCell && !isValidationStateNone) ? null : (
           <InputChevronIcon
             onClick={() => {
               if (!props.isDisabled) {
@@ -313,6 +333,7 @@ const _BaseDropdownInputTrigger = (
           />
         )
       }
+      {...(isInsideTableEditableCell ? tableInputProps : undefined)}
     />
   );
 };